Appdesktop/requirements.txt
PySide6>=6.5.0
requests>=2.28.0
paho-mqtt>=2.0.0
